# dev/gin/swagger/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "我的 gin swagger API 描述",
    "title": "我的 gin swagger API",
    "contact": {},
    "version": "1.0"
  },
  "paths": {
    "/api/get/{id}": {
      "get": {
        "security": [
          {
            "ApiToken": []
          }
        ],
        "tags": [
          "组1"
        ],
        "summary": "一个 get 请求",
        "parameters": [
          {
            "type": "string",
            "description": "ID 信息",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "description": "第几页数据",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "每页多少条数据",
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/main.ReturnMessage"
            }
          }
        }
      }
    },
    "/api/post/formdata": {
      "post": {
        "tags": [
          "组2"
        ],
        "summary": "需要 formdata 参数",
        "parameters": [
          {
            "type": "string",
            "description": "姓名",
            "name": "name",
            "in": "formData",
            "required": true
          },
          {
            "type": "file",
            "description": "需要上传的文件",
            "name": "upload_file",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/main.ReturnMessage"
            }
          }
        }
      }
    },
    "/api/post/{id}": {
      "post": {
        "tags": [
          "组2"
        ],
        "summary": "一个 post 请求",
        "parameters": [
          {
            "type": "string",
            "description": "ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "JSON数据",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/main.PostReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/main.ReturnMessage"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "main.PostReq": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "main.ReturnMessage": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "msg": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "ApiToken": {
      "type": "apiKey",
      "name": "X-Token",
      "in": "header"
    }
  }
}